Rename pipeline to task
