+++ apps/helm/taskmanager/chart/values.yaml
@@ -1,7 +1,7 @@
 ### Variables
 
 # Name
-name: pipelinevisualizer
+name: taskmanager
 
 # Logs
 logs:
@@ -31,7 +31,7 @@ database:
   # Address
   addresses:
     # Master
-    master: "mongodb-headless.platform.svc.cluster.local"
+    master: "mongodb-arbiter-headless.platform.svc.cluster.local"
     # Slave
     slave: "mongodb-headless.platform.svc.cluster.local"
   # Username
@@ -51,3 +51,8 @@ cache:
   port: 6379
   # Password
   password: "megasecret"
+
+# Broker
+broker:
+  # Address
+  address: "kafka.platform.svc.cluster.local:9092"
