+++ apps/helm/taskvisualizer/chart/Chart.yaml
@@ -1,5 +1,5 @@
 apiVersion: v2
-name: pipelinevisualizer
+name: taskvisualizer
 description: A Helm chart for Kubernetes
 
 # A chart can be either an 'application' or a 'library' chart.
