+++ apps/src/build_push_images.sh
@@ -90,37 +90,37 @@ pipelinevisualizerImageName="${containerRegistry}/${containerRegistryUsername}/$
 ### Build & Push ###
 ####################
 
-# organizationgateway
-docker build \
-  --platform "linux/${platform}" \
-  --tag "${organizationgatewayImageName}" \
-  --build-arg="APP_NAME=${organizationgateway}" \
-  "."
-docker push "${organizationgatewayImageName}"
+# # organizationgateway
+# docker build \
+#   --platform "linux/${platform}" \
+#   --tag "${organizationgatewayImageName}" \
+#   --build-arg="APP_NAME=${organizationgateway}" \
+#   "."
+# docker push "${organizationgatewayImageName}"
 
-# organizationmanager
-docker build \
-  --platform "linux/${platform}" \
-  --tag "${organizationmanagerImageName}" \
-  --build-arg="APP_NAME=${organizationmanager}" \
-  "."
-docker push "${organizationmanagerImageName}"
+# # organizationmanager
+# docker build \
+#   --platform "linux/${platform}" \
+#   --tag "${organizationmanagerImageName}" \
+#   --build-arg="APP_NAME=${organizationmanager}" \
+#   "."
+# docker push "${organizationmanagerImageName}"
 
-# jobgateway
-docker build \
-  --platform "linux/${platform}" \
-  --tag "${jobgatewayImageName}" \
-  --build-arg="APP_NAME=${jobgateway}" \
-  "."
-docker push "${jobgatewayImageName}"
+# # jobgateway
+# docker build \
+#   --platform "linux/${platform}" \
+#   --tag "${jobgatewayImageName}" \
+#   --build-arg="APP_NAME=${jobgateway}" \
+#   "."
+# docker push "${jobgatewayImageName}"
 
-# jobmanager
-docker build \
-  --platform "linux/${platform}" \
-  --tag "${jobmanagerImageName}" \
-  --build-arg="APP_NAME=${jobmanager}" \
-  "."
-docker push "${jobmanagerImageName}"
+# # jobmanager
+# docker build \
+#   --platform "linux/${platform}" \
+#   --tag "${jobmanagerImageName}" \
+#   --build-arg="APP_NAME=${jobmanager}" \
+#   "."
+# docker push "${jobmanagerImageName}"
 
 # jobvisualizer
 docker build \
@@ -130,26 +130,26 @@ docker build \
   "."
 docker push "${jobvisualizerImageName}"
 
-# pipelinegateway
-docker build \
-  --platform "linux/${platform}" \
-  --tag "${pipelinegatewayImageName}" \
-  --build-arg="APP_NAME=${pipelinegateway}" \
-  "."
-docker push "${pipelinegatewayImageName}"
+# # pipelinegateway
+# docker build \
+#   --platform "linux/${platform}" \
+#   --tag "${pipelinegatewayImageName}" \
+#   --build-arg="APP_NAME=${pipelinegateway}" \
+#   "."
+# docker push "${pipelinegatewayImageName}"
 
-# pipelinemanager
-docker build \
-  --platform "linux/${platform}" \
-  --tag "${pipelinemanagerImageName}" \
-  --build-arg="APP_NAME=${pipelinemanager}" \
-  "."
-docker push "${pipelinemanagerImageName}"
+# # pipelinemanager
+# docker build \
+#   --platform "linux/${platform}" \
+#   --tag "${pipelinemanagerImageName}" \
+#   --build-arg="APP_NAME=${pipelinemanager}" \
+#   "."
+# docker push "${pipelinemanagerImageName}"
 
-# pipelinevisualizer
-docker build \
-  --platform "linux/${platform}" \
-  --tag "${pipelinevisualizerImageName}" \
-  --build-arg="APP_NAME=${pipelinevisualizer}" \
-  "."
-docker push "${pipelinevisualizerImageName}"
+# # pipelinevisualizer
+# docker build \
+#   --platform "linux/${platform}" \
+#   --tag "${pipelinevisualizerImageName}" \
+#   --build-arg="APP_NAME=${pipelinevisualizer}" \
+#   "."
+# docker push "${pipelinevisualizerImageName}"
