+++ apps/src/taskmanager/requirements.txt
